Renaming MessagingApp to AppMessaging.
